requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
anthropic>=0.25.0
python-telegram-bot>=20.0
pyyaml>=6.0
